+++ CHANGELOG.md
@@ -33,6 +33,8 @@
 - Added OpenGL program representation (``src/gui/program.h``, ``src/gui/program.cpp``)
 - Fixed argument parsing bug
 - Fixed potential memory leak in ``src/gui/shader`` (``char *message = (char *)malloc(length);``)
+- Fixed OpenGL shader representation (``src/gui/shader.h``, ``src/gui/shader.cpp``)
+- Fixed OpenGL program representation (``src/gui/program.h``, ``src/gui/program.cpp``)
 
 2.0.1 - 2022-11-22
 ------------------